Big refactoring
