+++ Cargo.toml
@@ -16,7 +16,7 @@ build = "build.rs"
 [dependencies]
 clap = "3.0.0-beta.2"
 comfy-table = "1.1"
-orthanc = "0.5.2"
+orthanc = "0.6"
 serde_json = "1.0"
 serde_yaml = "0.8"
 
@@ -38,6 +38,10 @@ clap_generate = "3.0.0-beta.2"
 [[bin]]
 name = "orthanc"
 path = "src/main.rs"
+
+[lib]
+name = "orthanc_cli"
+path = "src/lib.rs"
 
 [[test]]
 name = "integration"
