+++ tests/data/patient_show.stdout
@@ -3,4 +3,4 @@
  PatientName         Patient 2
  PatientSex          M
  PatientBirthDate    19790101
- Number of studies   1
+ Number of Studies   1
