+++ tests/data/patient_show_anonymized_no_config.stdout
@@ -3,4 +3,4 @@
  PatientName         Anonymized502
  PatientSex
  PatientBirthDate
- Number of studies   1
+ Number of Studies   1
