+++ tests/data/patient_show_anonymized_with_config.stdout
@@ -3,4 +3,4 @@
  PatientName         Morty Smith
  PatientSex          M
  PatientBirthDate    19790101
- Number of studies   1
+ Number of Studies   1
