+++ tests/data/patient_show_modified.stdout
@@ -3,4 +3,4 @@
  PatientName         Rick Sanchez
  PatientSex          undefined
  PatientBirthDate    undefined
- Number of studies   1
+ Number of Studies   1
