+++ tests/data/series_show.stdout
@@ -5,4 +5,4 @@
  SeriesDescription     Series 1
  Modality              MR
  BodyPartExamined      KNEE
- Number of instances   1
+ Number of Instances   1
