+++ tests/data/series_show_anonymized_no_config.stdout
@@ -5,4 +5,4 @@
  SeriesDescription     undefined
  Modality              MR
  BodyPartExamined      KNEE
- Number of instances   1
+ Number of Instances   1
