+++ tests/data/series_show_anonymized_with_config.stdout
@@ -5,4 +5,4 @@
  SeriesDescription     Series 1
  Modality              MR
  BodyPartExamined      PINKY
- Number of instances   1
+ Number of Instances   1
