+++ tests/data/series_show_modified.stdout
@@ -5,4 +5,4 @@
  SeriesDescription     foobar-bazqux
  Modality              undefined
  BodyPartExamined      undefined
- Number of instances   1
+ Number of Instances   1
