+++ tests/data/study_show.stdout
@@ -7,4 +7,4 @@
  StudyDescription   Study 1
  StudyDate          20110101
  StudyTime          140606
- Number of series   2
+ Number of Series   2
