+++ tests/data/study_show_anonymized_no_config.stdout
@@ -7,4 +7,4 @@
  StudyDescription   undefined
  StudyDate
  StudyTime
- Number of series   2
+ Number of Series   2
