+++ tests/data/study_show_anonymized_with_config.stdout
@@ -7,4 +7,4 @@
  StudyDescription   undefined
  StudyDate          20110101
  StudyTime          140606
- Number of series   2
+ Number of Series   2
