+++ tests/data/study_show_modified.stdout
@@ -7,4 +7,4 @@
  StudyDescription   foobar
  StudyDate          undefined
  StudyTime          undefined
- Number of series   3
+ Number of Series   3
